resize necesario en las imagenes

+++ categories/card.php
@@ -7,8 +7,19 @@
       </div>
     </div>
     <div class="content">
-      <a class="button is-link" href="/categories/update.php?id=<?= $category['id'] ?>">Edit</a>
-      <a class="button is-danger" href="/categories/delete.php?id=<?= $category['id'] ?>">Delete</a>
+      <?php
+      if ($_SESSION['user_admin'] == 't') {
+      ?>
+        <a class="button is-link" href="/categories/update.php?id=<?= $category['id'] ?>">Edit</a>
+        <a class="button is-danger" href="/categories/delete.php?id=<?= $category['id'] ?>">Delete</a>
+      <?php
+      } elseif ($_SESSION['user_admin'] == 'f') {
+      ?>
+        <a class="button is-link" href="">Carrito</a>
+        <a class="button is-danger" href="">Ver</a>
+      <?php
+      }
+      ?>
     </div>
   </div>
 </div>
